backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
aiofiles>=23.2.1
pydantic>=2.7.0
pydantic-settings>=2.2.1
openai>=1.0.0
llama-index-core>=0.14.0
llama-index-readers-file>=0.6.0
llama-index-embeddings-openai>=0.6.0
pypdf>=4.0.0
python-pptx>=0.6.23
